Move number chooser classes to package component
