+++ app/src/main/java/com/armpatch/android/workouttracker/component/NumberChooser.java
@@ -1,4 +1,4 @@
-package com.armpatch.android.workouttracker;
+package com.armpatch.android.workouttracker.component;
 
 import android.view.View;
 import android.view.inputmethod.EditorInfo;
